--- TypeScript-Node-Starter/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "hallowtsserver",
-  "version": "1.0.0",
-  "description": "",
-  "scripts": {
-    "start": "npm run serve",
-    "build": "npm run build-ts && npm run tslint",
-    "serve": "node dist/server.js",
-    "watch-node": "nodemon dist/server.js",
-    "watch": "concurrently -k -p \"[{name}]\" -n \"Sass,TypeScript,Node\" -c \"yellow.bold,cyan.bold,green.bold\" \"npm run watch-ts\" \"npm run watch-node\"",
-    "build-ts": "tsc",
-    "watch-ts": "tsc -w",
-    "tslint": "tslint -c tslint.json -p tsconfig.json",
-    "debug": "npm run build && npm run watch-debug",
-    "serve-debug": "nodemon --inspect dist/server.js",
-    "watch-debug": "concurrently -k -p \"[{name}]\" -n \"TypeScript,Node\" -c \"yellow.bold,cyan.bold,green.bold\" \"npm run watch-ts\" \"npm run serve-debug\""
-  },
-  "dependencies": {
-    "async": "^2.6.0",
-    "mongoose": "^4.13.11"
-  },
-  "devDependencies": {
-    "@types/async": "^2.0.45",
-    "@types/mongodb": "^3.0.5",
-    "@types/mongoose": "^4.7.34",
-    "@types/node": "^9.4.6",
-    "@types/shelljs": "^0.7.8",
-    "chai": "^4.1.2",
-    "concurrently": "^3.5.1",
-    "nodemon": "^1.13.0",
-    "shelljs": "^0.8.1",
-    "ts-node": "^5.0.0",
-    "tslint": "^5.9.1",
-    "typescript": "^2.7.2"
-  }
-}
