--- TypeScript-Node-Starter/src/BusinessClasses/MapController.ts
@@ -1,4 +1,0 @@
-// @flow
-export default class MapController {
-
-};
