--- TypeScript-Node-Starter/src/app.ts
@@ -1,11 +1,0 @@
-
-/*
-import * as homeController from "./controllers/home";
-import * as userController from "./controllers/user";
-import * as apiController from "./controllers/api";
-import * as contactController from "./controllers/contact";
-*/
-
-const app = {};
-
-export default app;
