--- TypeScript-Node-Starter/src/data/worldMap.json
@@ -1,4 +1,0 @@
-[
-    [{"name": "0-0.json"}, {"name": "0-0.json"}],
-    [{"name": "0-0.json"}, {"name": "0-0.json"}]
-]
