--- TypeScript-Node-Starter/src/server.ts
@@ -1,19 +1,0 @@
-import app from "./app";
-
-/**
- * Start Express server.
- */
-
-/*
-const server = app.listen(app.get("port"), () => {
-  console.log(
-    "  App is running at http://localhost:%d in %s mode",
-    app.get("port"),
-    app.get("env")
-  );
-  console.log("  Press CTRL-C to stop\n");
-});*/
-
-const server = {};
-
-export default server;
