--- server/.flowconfig
@@ -1,11 +1,0 @@
-[ignore]
-
-[include]
-
-[libs]
-
-[lints]
-
-[options]
-
-[strict]
