+++ server/.travis.yml
