--- server/BusinessClasses/MapController.js
@@ -1,4 +1,0 @@
-// @flow
-export default class MapController {
-
-};
