+++ server/README.md
@@ -1,0 +1,2 @@
+# How to launch 
+npm run watch
