+++ server/package.json
@@ -1,19 +1,38 @@
 {
-  "name": "server",
+  "name": "hallowtsserver",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "npm run serve",
+    "build": "npm run build-ts && npm run tslint",
+    "serve": "node dist/server.js",
+    "watch-node": "nodemon dist/server.js",
+    "watch": "concurrently -k -p \"[{name}]\" -n \"TypeScript,Node\" -c \"yellow.bold,cyan.bold,green.bold\" \"npm run watch-ts\" \"npm run watch-node\"",
+    "build-ts": "tsc",
+    "watch-ts": "tsc -w",
+    "tslint": "tslint -c tslint.json -p tsconfig.json",
+    "debug": "npm run build && npm run watch-debug",
+    "serve-debug": "nodemon --inspect dist/server.js",
+    "watch-debug": "concurrently -k -p \"[{name}]\" -n \"TypeScript,Node\" -c \"yellow.bold,cyan.bold,green.bold\" \"npm run watch-ts\" \"npm run serve-debug\""
   },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
   "dependencies": {
-    "mongoose": "^5.2.3",
+    "@types/socket.io": "^1.4.36",
+    "async": "^2.6.0",
+    "mongoose": "^4.13.11",
     "socket.io": "^2.1.1"
   },
   "devDependencies": {
-    "flow-bin": "^0.76.0"
+    "@types/async": "^2.0.45",
+    "@types/mongodb": "^3.0.5",
+    "@types/mongoose": "^4.7.34",
+    "@types/node": "^9.4.6",
+    "@types/shelljs": "^0.7.8",
+    "chai": "^4.1.2",
+    "concurrently": "^3.5.1",
+    "nodemon": "^1.13.0",
+    "shelljs": "^0.8.1",
+    "ts-node": "^5.0.0",
+    "tslint": "^5.9.1",
+    "typescript": "^2.7.2"
   }
 }
