+++ server/src/BusinessClasses/MapController.ts
@@ -1,0 +1,18 @@
+// @flow
+export default class MapController {
+    worldMap: { name: string; }[][];
+    constructor() {
+        this.worldMap = [
+            [{name: "0-0.json"}, {name: "0-0.json"}],
+            [{name: "0-0.json"}, {name: "0-0.json"}]
+        ];
+    }
+
+    async movePlayer() {
+
+    }
+
+    async spawnPlayer() {
+
+    }
+}
