+++ server/src/BusinessClasses/State.ts
@@ -5,11 +5,6 @@ export default class State {
     MapController: MapController;
 
     constructor(){
-        // Load the real world map.
-        this.worldMap = [
-            [{"name": "0-0.json"}, {"name": "0-0.json"}],
-            [{"name": "0-0.json"}, {"name": "0-0.json"}]
-        ];
         this.MapController = new MapController();
     }
