+++ server/src/SocketHandler/Class/MapHandler.ts
